refactor: discountType 변환 로직 remote 패키지로 이동
